Fix 'shift count >= width of type' compiler warning in GC_SQRT_SIZE_MAX (fix commits 83231d0, 4e1a6f9)
* malloc.c (GC_SQRT_SIZE_MAX): Change type from unsigned to size_t.
* typd_mlc.c (GC_SQRT_SIZE_MAX): Likewise.

+++ typd_mlc.c
@@ -634,7 +634,7 @@ GC_API GC_ATTR_MALLOC void * GC_CALL
 # define GC_SIZE_MAX (~(size_t)0)
 #endif
 
-#define GC_SQRT_SIZE_MAX ((1U << (WORDSZ / 2)) - 1)
+#define GC_SQRT_SIZE_MAX ((((size_t)1) << (WORDSZ / 2)) - 1)
 
 GC_API GC_ATTR_MALLOC void * GC_CALL GC_calloc_explicitly_typed(size_t n,
                                                         size_t lb, GC_descr d)
